requirements.txt
Flask==0.10.1
gunicorn==19.6.0
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-firebase==1.2
pytz==2016.4
requests==2.10.0
six==1.10.0
twilio==5.4.0
Werkzeug==0.11.9
wheel==0.26.0
